#96 fix refactoring problems
