+++ api/server/app/de/htwg/zeta/server/model/modelValidator/validator/rules/metaModelDependent/NodeAttributeEnumTypes.scala
@@ -1,7 +1,7 @@
 package de.htwg.zeta.server.model.modelValidator.validator.rules.metaModelDependent
 
-import de.htwg.zeta.common.models.modelDefinitions.concept.elements.AttributeType.MEnum
-import de.htwg.zeta.common.models.modelDefinitions.concept.elements.AttributeValue.EnumValue
+import de.htwg.zeta.common.models.project.concept.elements.AttributeType.MEnum
+import de.htwg.zeta.common.models.project.concept.elements.AttributeValue.EnumValue
 import de.htwg.zeta.common.models.project.concept.Concept
 import de.htwg.zeta.common.models.project.instance.elements.NodeInstance
 import de.htwg.zeta.server.model.modelValidator.Util
